Refactor table-rendering code
The code still has problems. For example, if one of the fields (except
the last one) is very long it will look weird since reflow is not
applied to all table fields. This should be fix somehow - maybe by
allowing field-width specifiers, or by calculating the optimal width for
the field... But then, if I decide to implement "optimal width search
algorithm" the formatter will start getting a little bit too smart for
my taste. I don't want it to become too complex and a general purpose
text formatter.